added file upload/download, pagination

+++ controllers/admin.js
@@ -1,3 +1,5 @@
+const fileHelper = require('../util/file');
+
 const { validationResult } = require('express-validator/check');
 
 const Product = require('../models/product');
@@ -15,9 +17,26 @@ exports.getAddProduct = (req, res, next) => {
 
 exports.postAddProduct = (req, res, next) => {
     const title = req.body.title;
-    const imageUrl = req.body.imageUrl;
+    const image = req.file;
     const description = req.body.description;
     const price = req.body.price;
+    if (!image) {
+        return res.status(422).render('admin/edit-product', { 
+            pageTitle: 'Add Product', 
+            path: '/admin/add-product',
+            editing: false,
+            hasError: true,
+            product: {
+                title: title,
+                description: description,
+                price: price,
+            },
+            errorMessage: "Attached file is not an image",
+            validationErrors: [],
+        });
+    }
+
+
     const errors = validationResult(req);
 
     // catch input validation errors
@@ -29,7 +48,6 @@ exports.postAddProduct = (req, res, next) => {
             hasError: true,
             product: {
                 title: title,
-                imageUrl: imageUrl,
                 description: description,
                 price: price,
             },
@@ -37,6 +55,8 @@ exports.postAddProduct = (req, res, next) => {
             validationErrors: errors.array()
         });
     }
+    // stores image path in the file system
+    const imageUrl = image.path;
 
     const product = new Product({
         title: title,
@@ -89,7 +109,7 @@ exports.postEditProduct = (req, res, next) => {
     const prodId = req.body.productId;
     const updatedTitle = req.body.title;
     const updatedPrice = req.body.price;
-    const updatedImageUrl = req.body.imageUrl;
+    const image = req.file;
     const updatedDescription = req.body.description;
     const errors = validationResult(req);
 
@@ -101,7 +121,6 @@ exports.postEditProduct = (req, res, next) => {
             hasError: true,
             product: {
                 title: updatedTitle,
-                imageUrl: updatedImageUrl,
                 description: updatedDescription,
                 price: updatedPrice,
                 _id: prodId,
@@ -119,7 +138,10 @@ exports.postEditProduct = (req, res, next) => {
                 return res.redirect('/');
             }
             product.title = updatedTitle;
-            product.imageUrl = updatedImageUrl;
+            if (image) {
+                fileHelper.deleteFile(product.imageUrl);
+                product.imageUrl = image.path;
+            }
             product.description = updatedDescription;
             product.price = updatedPrice;
             return product.save().then(result => {
@@ -153,8 +175,17 @@ exports.getProducts = (req, res, next) => {
 
 exports.postDeleteProduct = (req, res, next) => {
     const prodId = req.body.productId;
-    Product
-        .deleteOne({_id: prodId, userId: req.user._id})
+    Product.findById(prodId)
+        .then(product => {
+            if (!product) {
+                return next(new Error('Product not found'));
+            }
+            // Deletes product image from file system
+            fileHelper.deleteFile(product.imageUrl);
+            // Deletes product data from db
+            return Product
+                .deleteOne({_id: prodId, userId: req.user._id});
+        })
         .then(() => {
             console.log('Deleted Product!');
             res.redirect('/admin/all-products');
